refactor: fix get token imports
