+++ src/pages/stats/features/date.js
@@ -1,4 +1,4 @@
 export function parseDate(dateStr) {
   const [date, month, year] = dateStr.trim().split('.');
-  return {date: +date, month: month - 1, year: +year};
+  return { date: +date, month: month - 1, year: +year };
 }
